ref(replay): move configuration resources card and remove old banners
